+++ test/packagebundler.jl
@@ -225,3 +225,44 @@ end
     end
     @test isfile(out)
 end
+
+@testset "path_filterer" begin
+    @testset "subdirectory" begin
+        dir = joinpath(@__DIR__, "fixtures", "ignorefiles")
+        pred = JuliaHub._PackageBundler.path_filterer(dir)
+        @test pred(joinpath(dir, "Pkg3", "README.md"))
+
+        @test pred(joinpath(dir, "Pkg3", "src", "bar"))
+        @test !pred(joinpath(dir, "Pkg3", "src", "foo"))
+        @test pred(joinpath(dir, "Pkg3", "src", "fooo"))
+
+        @test !pred(joinpath(dir, "Pkg3", "test", "bar"))
+        @test !pred(joinpath(dir, "Pkg3", "test", "foo"))
+        @test pred(joinpath(dir, "Pkg3", "test", "fooo", "test"))
+
+        # Note: even though the test/foo and test/bar directories are
+        # excluded, the predicate function does not return false if you
+        # check for file within the directories.
+        @test pred(joinpath(dir, "Pkg3", "test", "bar", "test"))
+        @test pred(joinpath(dir, "Pkg3", "test", "foo", "test"))
+    end
+    @testset "toplevel" begin
+        dir = joinpath(@__DIR__, "fixtures", "ignorefiles", "Pkg3")
+        pred = JuliaHub._PackageBundler.path_filterer(dir)
+        @test pred(joinpath(dir, "README.md"))
+
+        @test pred(joinpath(dir, "src", "bar"))
+        @test !pred(joinpath(dir, "src", "foo"))
+        @test pred(joinpath(dir, "src", "fooo"))
+
+        @test !pred(joinpath(dir, "test", "bar"))
+        @test !pred(joinpath(dir, "test", "foo"))
+        @test pred(joinpath(dir, "test", "fooo", "test"))
+
+        # Note: even though the test/foo and test/bar directories are
+        # excluded, the predicate function does not return false if you
+        # check for file within the directories.
+        @test pred(joinpath(dir, "test", "bar", "test"))
+        @test pred(joinpath(dir, "test", "foo", "test"))
+    end
+end