+++ test/utils.jl
@@ -101,3 +101,31 @@ end
         Dict("_id_missing" => "123e4567-e89b-12d3-a456-426614174000"), "id", UUIDs.UUID
     )
 end
+
+@testset "_max_appbundle_dir_size" begin
+    # We check here that the `.juliabundleignore` is honored by making
+    # sure that the calculated total file size of the Pkg3/ directory is
+    dir = joinpath(@__DIR__, "fixtures", "ignorefiles", "Pkg3")
+
+    appbundle_files = String[]
+    JuliaHub._walk_appbundle_files(dir) do filepath
+        push!(appbundle_files, relpath(filepath, dir))
+    end
+    @test sort(appbundle_files) == [
+        ".gitignore", ".juliabundleignore", "Project.toml", "README.md",
+        joinpath("src", "Pkg3.jl"), joinpath("src", "bar"), joinpath("src", "fooo"),
+        joinpath("test", "fooo", "test"), joinpath("test", "runtests.jl"),
+    ]
+
+    # The files that are not meant to be included in the /Pkg3/ bundle here are
+    # all 50 byte files. Should they should show up in the total size here.
+    #
+    # Note: on windows, the files may be checked out with different line endings,
+    # so the total file size may be slightly different.
+    sz, sz_is_low = JuliaHub._max_appbundle_dir_size(dir)
+    @test sz_is_low
+    @test sz in (Sys.iswindows() ? (405, 432) : (405,))
+
+    _, sz_low = JuliaHub._max_appbundle_dir_size(dir; maxsize=200)
+    @test !sz_low
+end